+++ examples/TBmodel_examples/README.md
@@ -15,4 +15,4 @@ graphene: tbmodel = {'label':'graphene_pythtb', 'delta':?, 't':?}
 haldane: tbmodel = {'label':'haldane_pythtb', 'delta':?, 't1':?, 't2':?, 'phi':?}
 kane_mele: tbmodel = {'label':'kane_mele_pythtb', 'delta':?, 't':?, 'soc':?, 'rashba':?}
 fu_kane_mele: tbmodel = {'label':'fu_kane_mele_pythtb', 't':?, 'soc':?, 'dt':?}
-See https://pythtb.readthedocs.io/en/latest/ or src/PAOFLOW/models/tb_models.py for details
+See https://pythtb.readthedocs.io/en/latest/ or src/PAOFLOW/models/models.py for details
