+++ src/PAOFLOW/projection/do_atwfc_proj.py
@@ -748,38 +748,73 @@ def read_QE_wfc(data_controller, ik, ispin):
     """
     arry, attr = data_controller.data_dicts()
     if attr['nspin'] == 1 or attr['nspin'] == 4:
-        wfcfile = 'wfc{0}.dat'.format(ik + 1)
+        wfcfile = 'wfc{0}'.format(ik + 1)
     elif attr['nspin'] == 2 and ispin == 0:
-        wfcfile = 'wfcdw{0}.dat'.format(ik + 1)
+        wfcfile = 'wfcdw{0}'.format(ik + 1)
     elif attr['nspin'] == 2 and ispin == 1:
-        wfcfile = 'wfcup{0}.dat'.format(ik + 1)
+        wfcfile = 'wfcup{0}'.format(ik + 1)
     else:
         print('no wfc file found')
 
-    with FortranFile(os.path.join(attr['fpath'], wfcfile), 'r') as f:
-        record = f.read_ints(np.int32)
-        assert len(record) == 11, 'something wrong reading fortran binary file'
+    try:
+        with FortranFile(os.path.join(attr['fpath'], wfcfile + '.dat'), 'r') as f:
+            record = f.read_ints(np.int32)
+            assert len(record) == 11, 'something wrong reading fortran binary file'
 
-        ik_ = record[0]
-        assert ik + 1 == ik_, 'wrong k-point in wfc file???'
+            ik_ = record[0]
+            assert ik + 1 == ik_, 'wrong k-point in wfc file???'
 
-        xk = np.frombuffer(record[1:7], np.float64)
-        ispin = record[7]
-        gamma_only = record[8] != 0
-        scalef = np.frombuffer(record[9:], np.float64)[0]
-        # print('ik =', ik, '  ispin =', ispin, '  gamma_only =', gamma_only, '  scalef =', scalef)
-        # print('xk =', xk)
+            xk = np.frombuffer(record[1:7], np.float64)
+            ispin = record[7]
+            gamma_only = record[8] != 0
+            scalef = np.frombuffer(record[9:], np.float64)[0]
+            # print('ik =', ik, '  ispin =', ispin, '  gamma_only =', gamma_only, '  scalef =', scalef)
+            # print('xk =', xk)
 
-        ngw, igwx, npol, nbnd = f.read_ints(np.int32)
-        # print('ngw =', ngw, '  igwx = ', igwx, '  npol = ', npol, '  nbnd = ', nbnd)
-        bg = f.read_reals(np.float64).reshape(3, 3, order='F')
-        # print('bg = ', bg)
-        mill = f.read_ints(np.int32).reshape(3, igwx, order='F')
-        # print('mill.shape = ', mill.shape)
+            ngw, igwx, npol, nbnd = f.read_ints(np.int32)
+            # print('ngw =', ngw, '  igwx = ', igwx, '  npol = ', npol, '  nbnd = ', nbnd)
+            bg = f.read_reals(np.float64).reshape(3, 3, order='F')
+            # print('bg = ', bg)
+            mill = f.read_ints(np.int32).reshape(3, igwx, order='F')
+            # print('mill.shape = ', mill.shape)
 
-        wfc = []
-        for i in range(nbnd):
-            wfc.append(f.read_reals(np.complex128))
+            wfc = []
+            for i in range(nbnd):
+                wfc.append(f.read_reals(np.complex128))
+
+    except:
+        try:
+            import h5py
+
+            with h5py.File(os.path.join(attr['fpath'], wfcfile + '.hdf5'), 'r') as h5:
+                ik_ = int(h5.attrs['ik'])
+                assert ik + 1 == ik_, 'wrong k-point in wfc file???'
+
+                xk = np.array(h5.attrs['xk'])
+                ispin = int(h5.attrs['ispin'])
+                gamma_only = 'TRUE' in h5.attrs['gamma_only'].decode()
+                scalef = float(h5.attrs['scale_factor'])
+                igwx = int(h5.attrs['igwx'])
+                npol = int(h5.attrs['npol'])
+                nbnd = int(h5.attrs['nbnd'])
+
+                miller_dataset = h5['MillerIndices']
+                bg = np.transpose(
+                    np.array(
+                        [
+                            miller_dataset.attrs['bg1'],
+                            miller_dataset.attrs['bg2'],
+                            miller_dataset.attrs['bg3'],
+                        ],
+                        dtype=np.float64,
+                    )
+                )
+                mill = np.transpose(np.array(miller_dataset, dtype=np.int64))
+
+                evc = np.array(h5['evc'], dtype=np.float64)
+                wfc = evc[:, 0::2] + 1j * evc[:, 1::2]
+        except:
+            raise Exception('no wfc file found')
 
     # compute overlap
     ovp = np.zeros((nbnd, nbnd), dtype=complex)
